requirements.txt
flake8==3.8.4
vulture==2.1
pylint-runner==0.6.0
bandit==1.6.2
psycopg2-binary==2.8.6